[client] [dashboard] [accounts_head] :: Move services to Accounts_Management level
